Update slider.less
I have one very low-contrast monitor at home and with it, I'm having a difficult time seeing the slider fill and handle. The handle is the active button, so I believe it should take on the primary color of the theme. This causes a little conflict with the range numbers being the same color, so I've changed the selected range color to instead include a rounded border with the primary-color. Lastly, I've changed the fill color to the greyDark color in the theme for better contrast.